fix: add support for dx/dy transform to Grid marks
